pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Open-source voice agent that can join Teams/SIP calls, gather user intent in <=10 exchanges, and produce a structured AI-ready prompt."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "playwright>=1.46.0",
  "faster-whisper>=1.0.3",
  "webrtcvad>=2.0.10",
  "sounddevice>=0.4.7",
  "numpy>=2.0.0",
  "llama-cpp-python>=0.2.90",
  "jinja2>=3.1.4"
]

[project.scripts]
voice-agent = "voice_agent.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
